+++ rector.php
@@ -3,6 +3,7 @@
 use Rector\Caching\ValueObject\Storage\FileCacheStorage;
 use Rector\Config\RectorConfig;
 use Rector\Php81\Rector\Array_\FirstClassCallableRector;
+use Rector\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector;
 use RectorLaravel\Rector\Class_\UnifyModelDatesWithCastsRector;
 use RectorLaravel\Rector\FuncCall\RemoveDumpDataDeadCodeRector;
 use RectorLaravel\Rector\MethodCall\EloquentWhereRelationTypeHintingParameterRector;
@@ -37,10 +38,23 @@ return RectorConfig::configure()
     ->withPhpSets(php82: true)
 
     ->withSkip([
+        // Je trouve la lecture plus difficile avec cette syntaxe, donc je désactive (PHP 7.4/8.1)
+        // ArraySpreadInsteadOfArrayMergeRector::class,
+
+        // Ne pas changer les closure et Arrow Function en Static
+        // StaticClosureRector::class,
+        // StaticArrowFunctionRector::class,
+
         // Désactivation de cette règle car elle
         // transforme :     array_map('intval',
         // en :             array_map(intval(...),
         FirstClassCallableRector::class,
+
+        // Cet attribut natif PHP n'est pas très utile ;
+        // mieux vaux se baser sur de l'analyse statique
+        // En plus, lors de la rédaction de ce message,
+        // la règle fonctionne mal, elle est buguée...
+        AddOverrideAttributeToOverriddenMethodsRector::class,
     ])
     ->withRules([
         EloquentWhereRelationTypeHintingParameterRector::class,